feat: add module browser-page,browser-scroll

+++ packages/core/src/core.ts
@@ -1,4 +1,4 @@
-import type { AnyFun } from "@tracing/shared";
+import { isBool, type AnyFun } from "@tracing/shared";
 import type { TracingCoreConfig } from "./config";
 import type { Store } from "./store";
 
@@ -8,8 +8,6 @@ import { createInitConfig } from "./config";
 import { createLogger } from "./logger";
 
 export class TracingCore {
-  [key: string]: any;
-
   private initialized = false;
   private destroyed = false;
 
@@ -22,11 +20,11 @@ export class TracingCore {
 
   constructor(config: Partial<TracingCoreConfig>) {
     this.config = createInitConfig(config);
-    this.logger = createLogger("core", this.config);
+    this.logger = createLogger("core", !!this.config.sendLog);
 
     this.pluginDriver = PluginDriver(this.config.plugins.flat(), this.config, this.logger);
 
-    this.pluginDriver.hookSequentialSync("setup", [this.config]);
+    this.header = createStore("header", {});
   }
 
   public init() {
@@ -35,10 +33,11 @@ export class TracingCore {
       return;
     }
 
-    this.header = createStore("header", {});
+    this.pluginDriver.hookSequentialSync("init", [this]);
 
     this.initialized = true;
-    this.effects = this.pluginDriver.hookSequentialSync("init", [this]);
+
+    this.effects = this.pluginDriver.hookSequentialSync("setup", [this]);
   }
 
   public report(event: string, record: Record<string, any>, meta?: unknown) {
@@ -58,6 +57,14 @@ export class TracingCore {
   private build(event: string, record: Record<string, any>, meta?: unknown) {
     const build = this.pluginDriver.hookChainMergeSync("build", [event, record, meta], {});
 
+    if (__DEV__ && this.config.sendLog) {
+      if (isBool(this.config.sendLog)) {
+        console.log(JSON.stringify(build, null, 2));
+      } else {
+        this.config.sendLog(event, build);
+      }
+    }
+
     this.send(event, build);
   }
 
@@ -65,23 +72,26 @@ export class TracingCore {
     const isBail = this.pluginDriver.hookBail("beforeSend", [event, build]);
     if (isBail === false) return;
 
-    if (__DEV__ && this.config.isLogger && this.config.sendLog) {
-      console.log(JSON.stringify(build, null, 2));
-    }
-
     this.pluginDriver.hookBail("send", [event, build]);
   }
 
   public async destroy() {
-    await Promise.all(this.effects.map(effect => effect()));
-    this.effects.length = 0;
+    if (this.destroyed) {
+      __DEV__ && this.logger.warn('"TracingCore" has been destroyed. Please do not retry');
+      return;
+    }
 
     await this.pluginDriver.hookParallel("beforeDestroy", [this]);
 
+    await Promise.all(this.effects.map(effect => effect()));
+
     this.pluginDriver.hookSequentialSync("destroy", [this]);
 
-    this.header.clear();
     this.pluginDriver.destroyPluginDriver();
+
+    this.header.clear();
+    this.effects.length = 0;
+
     this.destroyed = true;
   }
 }
